refactor: extract duplication from CachePolicy implementations
