+++ lib/config_cat/cache_policy/auto.ex
@@ -2,6 +2,7 @@ defmodule ConfigCat.CachePolicy.Auto do
   use GenServer
 
   alias ConfigCat.CachePolicy
+  alias ConfigCat.CachePolicy.Helpers
 
   defstruct poll_interval_seconds: 60, mode: "a"
 
@@ -13,25 +14,8 @@ defmodule ConfigCat.CachePolicy.Auto do
   end
 
   def start_link(options) do
-    {name, options} = Keyword.pop!(options, :name)
-
-    policy_options =
-      options
-      |> Keyword.fetch!(:cache_policy)
-      |> Map.from_struct()
-      |> Map.drop([:mode])
-
-    initial_state =
-      default_options()
-      |> Keyword.merge(options)
-      |> Keyword.take([:cache, :cache_key, :fetcher, :fetcher_id])
-      |> Enum.into(%{})
-      |> Map.merge(policy_options)
-
-    GenServer.start_link(__MODULE__, initial_state, name: name)
+    Helpers.start_link(__MODULE__, options)
   end
-
-  defp default_options, do: [fetcher: ConfigCat.CacheControlConfigFetcher]
 
   @impl GenServer
   def init(state) do
@@ -60,12 +44,12 @@ defmodule ConfigCat.CachePolicy.Auto do
 
   @impl GenServer
   def handle_call(:get, _from, state) do
-    cached_config(state)
+    {:reply, Helpers.cached_config(state), state}
   end
 
   @impl GenServer
   def handle_call(:force_refresh, _from, state) do
-    case refresh(state) do
+    case Helpers.refresh_config(state) do
       :ok ->
         {:reply, :ok, state}
 
@@ -74,35 +58,10 @@ defmodule ConfigCat.CachePolicy.Auto do
     end
   end
 
-  defp cached_config(state) do
-    cache = Map.fetch!(state, :cache)
-    cache_key = Map.fetch!(state, :cache_key)
-
-    {:reply, cache.get(cache_key), state}
-  end
-
   defp polled_refresh(%{poll_interval_seconds: seconds} = state) do
-    refresh(state)
+    Helpers.refresh_config(state)
     Process.send_after(self(), :polled_refresh, seconds * 1000)
 
     {:noreply, state}
-  end
-
-  defp refresh(state) do
-    fetcher = Map.fetch!(state, :fetcher)
-    fetcher_id = Map.fetch!(state, :fetcher_id)
-
-    case fetcher.fetch(fetcher_id) do
-      {:ok, :unchanged} ->
-        :ok
-
-      {:ok, config} ->
-        cache = Map.fetch!(state, :cache)
-        cache_key = Map.fetch!(state, :cache_key)
-        cache.set(cache_key, config)
-
-      error ->
-        error
-    end
   end
 end
