+++ lib/config_cat/cache_policy/lazy.ex
@@ -2,6 +2,7 @@ defmodule ConfigCat.CachePolicy.Lazy do
   use GenServer
 
   alias ConfigCat.CachePolicy
+  alias ConfigCat.CachePolicy.Helpers
 
   @enforce_keys [:cache_expiry_seconds]
   defstruct [:cache_expiry_seconds, mode: "l"]
@@ -13,26 +14,8 @@ defmodule ConfigCat.CachePolicy.Lazy do
   end
 
   def start_link(options) do
-    {name, options} = Keyword.pop!(options, :name)
-
-    policy_options =
-      options
-      |> Keyword.fetch!(:cache_policy)
-      |> Map.from_struct()
-      |> Map.drop([:mode])
-
-    initial_state =
-      default_options()
-      |> Keyword.merge(options)
-      |> Keyword.take([:cache, :cache_key, :fetcher, :fetcher_id])
-      |> Enum.into(%{})
-      |> Map.merge(policy_options)
-      |> Map.put(:last_update, nil)
-
-    GenServer.start_link(__MODULE__, initial_state, name: name)
+    Helpers.start_link(__MODULE__, options, %{last_update: nil})
   end
-
-  defp default_options, do: [fetcher: ConfigCat.CacheControlConfigFetcher]
 
   @impl GenServer
   def init(state) do
@@ -52,13 +35,13 @@ defmodule ConfigCat.CachePolicy.Lazy do
   @impl GenServer
   def handle_call(:get, _from, state) do
     with {:ok, new_state} <- maybe_refresh(state) do
-      cached_config(new_state)
+      {:reply, Helpers.cached_config(new_state), new_state}
     end
   end
 
   @impl GenServer
   def handle_call(:force_refresh, _from, state) do
-    case refresh_and_record(state) do
+    case refresh(state) do
       {:ok, new_state} ->
         {:reply, :ok, new_state}
 
@@ -67,16 +50,9 @@ defmodule ConfigCat.CachePolicy.Lazy do
     end
   end
 
-  defp cached_config(state) do
-    cache = Map.fetch!(state, :cache)
-    cache_key = Map.fetch!(state, :cache_key)
-
-    {:reply, cache.get(cache_key), state}
-  end
-
   defp maybe_refresh(state) do
     if needs_fetch?(state) do
-      refresh_and_record(state)
+      refresh(state)
     else
       {:ok, state}
     end
@@ -91,29 +67,10 @@ defmodule ConfigCat.CachePolicy.Lazy do
       |> DateTime.compare(now())
   end
 
-  defp refresh_and_record(state) do
-    case refresh(state) do
+  defp refresh(state) do
+    case Helpers.refresh_config(state) do
       :ok -> {:ok, %{state | last_update: now()}}
       error -> error
-    end
-  end
-
-  defp refresh(state) do
-    fetcher = Map.fetch!(state, :fetcher)
-    fetcher_id = Map.fetch!(state, :fetcher_id)
-
-    case fetcher.fetch(fetcher_id) do
-      {:ok, :unchanged} ->
-        :ok
-
-      {:ok, config} ->
-        cache = Map.fetch!(state, :cache)
-        cache_key = Map.fetch!(state, :cache_key)
-        cache.set(cache_key, config)
-        :ok
-
-      error ->
-        error
     end
   end
